pyproject.toml
[project]
name = "ai-doc-gen"
version = "1.2.0"
description = "AI-powered code documentation generator that analyzes repositories and creates comprehensive documentation"
authors = [{ name = "Milad Noroozi", email = "norooziosos@gmail.com" }]
requires-python = ">=3.13,<3.14"
dependencies = [
    "pydantic==2.12.4",
    "jinja2==3.1.6",
    "ujson==5.11.0",
    "pydantic-ai==1.22.0",
    "nest-asyncio==1.6.0",
    "python-gitlab==7.0.0",
    "gitpython==3.1.45",
    "logfire==4.15.1",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.1.1",
    "opentelemetry-instrumentation-httpx==0.58b0",
    "atlassian-python-api>=3.41.0",  # Bitbucket Server support
    "litellm>=1.55.0",  # Unified LLM provider interface
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py313"
fix = true
src = ["src"]

[tool.ruff.analyze]
detect-string-imports = true
direction = "dependencies"

[tool.ruff.format]
docstring-code-format = true


[dependency-groups]
dev = [
    "ipython>=9.6.0",
    "ruff>=0.14.0",
]

[project.scripts]
ai-doc-gen = "src.main:cli_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.sources]
"src" = "src"
